Moved the misbehaving test kernels inside the test file
