+++ cpp/src/arrow/compute/kernels/scalar_misbehave_test.cc
@@ -15,26 +15,72 @@
 // specific language governing permissions and limitations
 // under the License.
 
+#include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 #include "arrow/array/concatenate.h"
 #include "arrow/compute/api_scalar.h"
 #include "arrow/testing/gtest_util.h"
+#include "arrow/util/logging.h"
 
 namespace arrow {
 namespace compute {
-std::shared_ptr<const ScalarFunction> CreateScalarMisbehaveFunction();
 
+namespace {
+
+template <typename AllocateMem>
+struct ScalarMisbehaveExec {
+    static Status Exec(KernelContext* ctx, const ExecBatch& batch, Datum* out) {
+      // allocate a buffer even though we've promised not to
+      ARROW_ASSIGN_OR_RAISE(auto buffer, ctx->Allocate(64));
+      BufferVector buffers{nullptr, std::move(buffer)};
+      auto array = std::make_shared<ArrayData>(int64(), 8, std::move(buffers));
+      *out = array;
+      // return Status::NotImplemented("This kernel only exists for testing purposes");
+      // The function should return OK, otherwise the buffer check is not performed
+      return Status::OK();
+    }
+};
+}  // namespace internal
+
+void AddScalarMisbehaveKernels(const std::shared_ptr<ScalarFunction>& scalar_function) {
+//  ScalarKernel kernel({InputType(Type::FIXED_SIZE_BINARY)},
+//                      OutputType(ValueDescr(fixed_size_binary(2)))),
+//                      ScalarMisbehaveExec</*AllocateMem=*/std::false_type>::Exec);
+//  kernel.null_handling = NullHandling::COMPUTED_PREALLOCATE;
+//  kernel.mem_allocation = MemAllocation::PREALLOCATE; //is the default
+//  kernel.can_write_into_slices = true;
+
+  DCHECK_OK(scalar_function->AddKernel({InputType(Type::FIXED_SIZE_BINARY)},
+                                       OutputType(ValueDescr(fixed_size_binary(2))),
+                                       ScalarMisbehaveExec</*AllocateMem=*/std::false_type>::Exec));
+}
+
+const FunctionDoc misbehave_doc{
+        "Test kernel that does nothing but allocate memory "
+        "while it shouldn't",
+        "This Kernel only exists for testing purposes.\n"
+        "It allocates memory while it promised not to \n"
+        "(because of MemAllocation::PREALLOCATE).",
+        {}};
+
+std::shared_ptr<const ScalarFunction> CreateScalarMisbehaveFunction() {
+  auto func = std::make_shared<ScalarFunction>("scalar_misbehave", Arity::Unary(),
+                                               &misbehave_doc);
+  AddScalarMisbehaveKernels(func);
+  return func;
+}
 TEST(Misbehave, MisbehavingScalarKernel) {
   ExecContext ctx;
   auto func = CreateScalarMisbehaveFunction();
   Datum datum(ArrayFromJSON(fixed_size_binary(6), R"(["123456"])"));
   const std::vector<Datum> &args = {datum};
   const FunctionOptions *options = nullptr;
-  ASSERT_RAISES_WITH_MESSAGE(ExecutionError,
-                             "ExecutionError in Gandiva: "
-                             "Unauthorized memory allocations "
-                             "in function kernel",
-                             func->Execute(args, options, &ctx));
+  EXPECT_RAISES_WITH_MESSAGE_THAT(ExecutionError,
+                                  testing::HasSubstr(
+                                 "ExecutionError in Gandiva: "
+                                 "Unauthorized memory allocations "
+                                 "in function kernel"),
+                                 func->Execute(args, options, &ctx));
 }
 }  // namespace compute
 }  // namespace arrow
