+++ cpp/src/arrow/compute/kernels/vector_misbehave_test.cc
@@ -15,15 +15,55 @@
 // specific language governing permissions and limitations
 // under the License.
 
+#include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 #include "arrow/array/concatenate.h"
 #include "arrow/chunked_array.h"
 #include "arrow/compute/api_vector.h"
 #include "arrow/testing/gtest_util.h"
+#include "arrow/util/logging.h"
 
 namespace arrow {
 namespace compute {
-std::shared_ptr<const VectorFunction> CreateVectorMisbehaveFunction();
+namespace {
+
+template <typename AllocateMem>
+struct VectorMisbehaveExec {
+    static Status Exec(KernelContext* ctx, const ExecBatch& batch, Datum* out) {
+      // allocate new buffers even though we've promised not to
+      ARROW_ASSIGN_OR_RAISE(out->mutable_array()->buffers[0], ctx->AllocateBitmap(8));
+      ARROW_ASSIGN_OR_RAISE(out->mutable_array()->buffers[1], ctx->Allocate(64));
+      return Status::OK();
+    }
+};
+
+void AddVectorMisbehaveKernels(const std::shared_ptr<VectorFunction>& Vector_function) {
+  VectorKernel kernel({int32()}, int32(),
+                      VectorMisbehaveExec</*AllocateMem=*/std::false_type>::Exec);
+  kernel.null_handling = NullHandling::COMPUTED_PREALLOCATE;
+  kernel.mem_allocation = MemAllocation::PREALLOCATE;
+  kernel.can_write_into_slices = true;
+  kernel.can_execute_chunkwise = false;
+  kernel.output_chunked = false;
+
+  DCHECK_OK(Vector_function->AddKernel(std::move(kernel)));
+}
+}  // namespace
+
+const FunctionDoc misbehave_doc{
+        "Test kernel that does nothing but allocate memory "
+        "while it shouldn't",
+        "This Kernel only exists for testing purposes.\n"
+        "It allocates memory while it promised not to \n"
+        "(because of MemAllocation::PREALLOCATE).",
+        {}};
+
+std::shared_ptr<const VectorFunction> CreateVectorMisbehaveFunction() {
+  auto func = std::make_shared<VectorFunction>("vector_misbehave", Arity::Unary(),
+                                               &misbehave_doc);
+  AddVectorMisbehaveKernels(func);
+  return func;
+}
 
 TEST(Misbehave, MisbehavingVectorKernel) {
   ExecContext ctx;
@@ -31,11 +71,12 @@ TEST(Misbehave, MisbehavingVectorKernel) {
   Datum datum(ChunkedArray(ArrayVector{}, int32()));
   const std::vector<Datum> &args = {datum};
   const FunctionOptions *options = nullptr;
-  ASSERT_RAISES_WITH_MESSAGE(ExecutionError,
-                             "ExecutionError in Gandiva: "
-                             "Unauthorized memory allocations "
-                             "in function kernel",
-                             func->Execute(args, options, &ctx));
+  EXPECT_RAISES_WITH_MESSAGE_THAT(ExecutionError,
+                                  testing::HasSubstr(
+                                          "ExecutionError in Gandiva: "
+                                          "Unauthorized memory allocations "
+                                          "in function kernel"),
+                                  func->Execute(args, options, &ctx));
 }
 }  // namespace compute
 }  // namespace arrow
